recipe-backend/requirements.txt
Flask==2.1.0
requests==2.27.1
flask_cors==3.1.1
python-dotenv==0.19.2